requirements.txt
psutil
colorama
keyboard
pyautogui
unipath
mouse
pydirectinput
winshell
pypiwin32
Pillow
opencv-python
gitpython
pypresence
anonfile
playsound==1.2.2
PyQt5
pyfiglet